edit

+++ scripts/2017.4.sql
@@ -1491,7 +1491,7 @@ CREATE TABLE `project_view` (
 CREATE TABLE `bankinfo_log` (
   `id` int(11) NOT NULL AUTO_INCREMENT,
   `bankname` varchar(50) NOT NULL COMMENT '开户银行',
-  `bankcard` int(50) NOT NULL COMMENT '银行卡号',
+  `bankcard` bigint(30) NOT NULL COMMENT '银行卡号',
   `username` varchar(50) NOT NULL COMMENT '开户名',
   `position` varchar(150) NOT NULL COMMENT '开户行所在地',
   `bankbranch` varchar(150) NOT NULL COMMENT '开户行支行名',
